+++ src/variable/Constant.cpp
@@ -1,10 +1,10 @@
 #include "Constant.h"
 
-Constant::Constant(uint value, uint address) : Variable(std::to_string(value), address) {
+Constant::Constant(uint64_t value, uint address) : Variable(std::to_string(value), address) {
     this->value = value;
     this->initialize();
 }
 
-uint Constant::getValue() {
+uint64_t Constant::getValue() {
     return this->value;
 }
